backend/pyproject.toml
[project]
name = "gapyearcode-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "httpx>=0.28,<1.0",
  "pydantic-settings>=2.8,<3.0",
  "uvicorn[standard]>=0.34,<1.0",
]

[project.scripts]
app = "app.main:app"
